Implemented a work-around for broken uploads to original wired servers.
Update issue 5
It seems that the original wired server daemon assumes that the client sends the TRANSFER message as a single data frame over the wire. Whenever this is not the case, it simply cuts off the beginning of the data transmission and writes data starting from frame 2 to the target file. Ouch.
After being on a slower connection without a complex router between me and the server, I could finally track this down. Let's hope this works.